Move some functionality to s3 package
